(preliminary) handle read_ids in subsampling
A fix had been implemented to prevent a duplicate reads error for a type
of SE file, but introduced bugs for other types of PE files. Remove that
previous fix and just don't strip anything off the read ID for SE files.

Note: these are preliminary changes and must be tested